Add integrations with custom items from external plugins (part 1)
The integrations only support VirtualShop as it's the best option for creating system shops. If the custom items in the shop are from external plugins, they will be loaded directly from the external plugins. This frees the server admins from manually updating the items in VirtualShop. For exmaple, if the custom items are changed in the database of external plugins, they will be updated and reflected in the VirtualShop as soon as running `/virtualshop reload`